+++ handlers/facebook/facebook.go
@@ -171,6 +171,7 @@ type moPayload struct {
 					Ref    string `json:"ref"`
 					Source string `json:"source"`
 					Type   string `json:"type"`
+					AdID   string `json:"ad_id"`
 				} `json:"referral"`
 			} `json:"postback"`
 
@@ -297,6 +298,10 @@ func (h *handler) receiveEvent(ctx context.Context, channel courier.Channel, w h
 				extra[referrerIDKey] = msg.Postback.Referral.Ref
 				extra[sourceKey] = msg.Postback.Referral.Source
 				extra[typeKey] = msg.Postback.Referral.Type
+
+				if msg.Postback.Referral.AdID != "" {
+					extra[adIDKey] = msg.Postback.Referral.AdID
+				}
 			}
 
 			event = event.WithExtra(extra)
@@ -536,9 +541,14 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 			return status, err
 		}
 
-		req, _ := http.NewRequest(http.MethodPost, msgURL.String(), bytes.NewReader(jsonBody))
+		req, err := http.NewRequest(http.MethodPost, msgURL.String(), bytes.NewReader(jsonBody))
+
+		if err != nil {
+			return nil, err
+		}
 		req.Header.Set("Content-Type", "application/json")
 		req.Header.Set("Accept", "application/json")
+
 		rr, err := utils.MakeHTTPRequest(req)
 
 		// record our status and log
@@ -604,7 +614,7 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 	return status, nil
 }
 
-// ReceiveVerify handles Facebook's webhook verification callback
+// DescribeURN looks up URN metadata for new contacts
 func (h *handler) DescribeURN(ctx context.Context, channel courier.Channel, urn urns.URN) (map[string]string, error) {
 	// can't do anything with facebook refs, ignore them
 	if urn.IsFacebookRef() {