+++ handlers/highconnection/highconnection.go
@@ -3,6 +3,7 @@ package highconnection
 import (
 	"context"
 	"fmt"
+	"mime"
 	"net/http"
 	"net/url"
 	"time"
@@ -66,8 +67,14 @@ func (h *handler) receiveMessage(ctx context.Context, channel courier.Channel, w
 		return nil, handlers.WriteAndLogRequestError(ctx, h, channel, w, r, err)
 	}
 
+	// Hign connection URL encodes escapes ISO 8859 escape sequences
+	text, _ := url.QueryUnescape(form.Message)
+	// decode from ISO 8859
+	text = mime.BEncoding.Encode("ISO-8859-1", text)
+	text, _ = new(mime.WordDecoder).DecodeHeader(text)
+
 	// build our Message
-	msg := h.Backend().NewIncomingMsg(channel, urn, form.Message).WithReceivedOn(date.UTC())
+	msg := h.Backend().NewIncomingMsg(channel, urn, utils.CleanString(text)).WithReceivedOn(date.UTC())
 
 	// and finally write our message
 	return handlers.WriteMsgsAndResponse(ctx, h, []courier.Msg{msg}, w, r)
@@ -143,7 +150,11 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 		msgURL, _ := url.Parse(sendURL)
 		msgURL.RawQuery = form.Encode()
 
-		req, _ := http.NewRequest(http.MethodPost, msgURL.String(), nil)
+		req, err := http.NewRequest(http.MethodPost, msgURL.String(), nil)
+
+		if err != nil {
+			return nil, err
+		}
 		rr, err := utils.MakeHTTPRequest(req)
 
 		// record our status and log