+++ handlers/whatsapp/whatsapp.go
@@ -15,29 +15,46 @@ import (
 	"github.com/nyaruka/courier"
 	"github.com/nyaruka/courier/handlers"
 	"github.com/nyaruka/courier/utils"
+	"github.com/nyaruka/gocommon/rcache"
 	"github.com/nyaruka/gocommon/urns"
+	"github.com/patrickmn/go-cache"
 	"github.com/pkg/errors"
+	"github.com/sirupsen/logrus"
 )
 
 const (
 	configNamespace  = "fb_namespace"
 	configHSMSupport = "hsm_support"
+
+	d3AuthorizationKey = "D360-API-KEY"
+
+	channelTypeWa  = "WA"
+	channelTypeD3  = "D3"
+	channelTypeTXW = "TXW"
+
+	mediaCacheKeyPattern = "whatsapp_media_%s"
 )
 
 var (
 	retryParam = ""
 )
 
+var failedMediaCache *cache.Cache
+
 func init() {
-	courier.RegisterHandler(newHandler())
+	courier.RegisterHandler(newWAHandler(courier.ChannelType(channelTypeWa), "WhatsApp"))
+	courier.RegisterHandler(newWAHandler(courier.ChannelType(channelTypeD3), "360Dialog"))
+	courier.RegisterHandler(newWAHandler(courier.ChannelType(channelTypeTXW), "TextIt"))
+
+	failedMediaCache = cache.New(15*time.Minute, 15*time.Minute)
 }
 
 type handler struct {
 	handlers.BaseHandler
 }
 
-func newHandler() courier.ChannelHandler {
-	return &handler{handlers.NewBaseHandler(courier.ChannelType("WA"), "WhatsApp")}
+func newWAHandler(channelType courier.ChannelType, name string) courier.ChannelHandler {
+	return &handler{handlers.NewBaseHandler(channelType, name)}
 }
 
 // Initialize is called by the engine once everything is loaded
@@ -86,6 +103,10 @@ type eventPayload struct {
 			MimeType string `json:"mime_type" validate:"required"`
 			Sha256   string `json:"sha256"    validate:"required"`
 		} `json:"audio"`
+		Button *struct {
+			Payload string `json:"payload"`
+			Text    string `json:"text"    validate:"required"`
+		} `json:"button"`
 		Document *struct {
 			File     string `json:"file"      validate:"required"`
 			ID       string `json:"id"        validate:"required"`
@@ -173,6 +194,8 @@ func (h *handler) receiveEvent(ctx context.Context, channel courier.Channel, w h
 			text = msg.Text.Body
 		} else if msg.Type == "audio" && msg.Audio != nil {
 			mediaURL, err = resolveMediaURL(channel, msg.Audio.ID)
+		} else if msg.Type == "button" && msg.Button != nil {
+			text = msg.Button.Text
 		} else if msg.Type == "document" && msg.Document != nil {
 			text = msg.Document.Caption
 			mediaURL, err = resolveMediaURL(channel, msg.Document.ID)
@@ -270,8 +293,8 @@ func (h *handler) BuildDownloadMediaRequest(ctx context.Context, b courier.Backe
 
 	// set the access token as the authorization header
 	req, _ := http.NewRequest(http.MethodGet, attachmentURL, nil)
-	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", token))
 	req.Header.Set("User-Agent", utils.HTTPUserAgent)
+	setWhatsAppAuthHeader(&req.Header, channel)
 	return req, nil
 }
 
@@ -319,7 +342,8 @@ type mtTextPayload struct {
 }
 
 type mediaObject struct {
-	Link    string `json:"link" validate:"required"`
+	ID      string `json:"id,omitempty"`
+	Link    string `json:"link,omitempty"`
 	Caption string `json:"caption,omitempty"`
 }
 
@@ -424,56 +448,62 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 	if len(msg.Attachments()) > 0 {
 		for attachmentCount, attachment := range msg.Attachments() {
 
-			mimeType, s3url := handlers.SplitAttachment(attachment)
-
+			mimeType, mediaURL := handlers.SplitAttachment(attachment)
+			mediaID, mediaLogs, err := h.fetchMediaID(msg, mimeType, mediaURL)
+			if err != nil {
+				logrus.WithField("channel_uuid", msg.Channel().UUID().String()).WithError(err).Error("error while uploading media to whatsapp")
+			}
+			if err == nil && mediaID != "" {
+				mediaURL = ""
+			}
+			mediaPayload := &mediaObject{ID: mediaID, Link: mediaURL}
 			externalID := ""
 			if strings.HasPrefix(mimeType, "audio") {
 				payload := mtAudioPayload{
 					To:   msg.URN().Path(),
 					Type: "audio",
 				}
-				payload.Audio = &mediaObject{Link: s3url}
-				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
-
+				payload.Audio = mediaPayload
+				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, payload)
 			} else if strings.HasPrefix(mimeType, "application") {
 				payload := mtDocumentPayload{
 					To:   msg.URN().Path(),
 					Type: "document",
 				}
-
 				if attachmentCount == 0 {
-					payload.Document = &mediaObject{Link: s3url, Caption: msg.Text()}
-				} else {
-					payload.Document = &mediaObject{Link: s3url}
+					mediaPayload.Caption = msg.Text()
 				}
-				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
-
+				payload.Document = mediaPayload
+				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, payload)
 			} else if strings.HasPrefix(mimeType, "image") {
 				payload := mtImagePayload{
 					To:   msg.URN().Path(),
 					Type: "image",
 				}
 				if attachmentCount == 0 {
-					payload.Image = &mediaObject{Link: s3url, Caption: msg.Text()}
-				} else {
-					payload.Image = &mediaObject{Link: s3url}
+					mediaPayload.Caption = msg.Text()
 				}
-				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
+				payload.Image = mediaPayload
+				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, payload)
 			} else if strings.HasPrefix(mimeType, "video") {
 				payload := mtVideoPayload{
 					To:   msg.URN().Path(),
 					Type: "video",
 				}
 				if attachmentCount == 0 {
-					payload.Video = &mediaObject{Link: s3url, Caption: msg.Text()}
-				} else {
-					payload.Video = &mediaObject{Link: s3url}
+					mediaPayload.Caption = msg.Text()
 				}
-				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
+				payload.Video = mediaPayload
+				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, payload)
 			} else {
 				duration := time.Since(start)
 				err = fmt.Errorf("unknown attachment mime type: %s", mimeType)
 				logs = []*courier.ChannelLog{courier.NewChannelLogFromError("Error sending message", msg.Channel(), msg.ID(), duration, err)}
+			}
+
+			// add media logs to our status
+			for _, log := range mediaLogs {
+				status.AddLog(log)
 			}
 
 			// add logs to our status
@@ -519,7 +549,7 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 				for _, v := range templating.Variables {
 					payload.HSM.LocalizableParams = append(payload.HSM.LocalizableParams, LocalizableParam{Default: v})
 				}
-				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
+				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, payload)
 			} else {
 
 				payload := templatePayload{
@@ -538,7 +568,7 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 				}
 				payload.Template.Components = append(payload.Template.Components, *component)
 
-				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
+				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, payload)
 			}
 
 			// add logs to our status
@@ -558,7 +588,7 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 					Type: "text",
 				}
 				payload.Text.Body = part
-				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, token, payload)
+				wppID, externalID, logs, err = sendWhatsAppMsg(msg, sendPath, payload)
 
 				// add logs to our status
 				for _, log := range logs {
@@ -594,7 +624,82 @@ func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStat
 	return status, nil
 }
 
-func sendWhatsAppMsg(msg courier.Msg, sendPath *url.URL, token string, payload interface{}) (string, string, []*courier.ChannelLog, error) {
+// fetchMediaID tries to fetch the id for the uploaded media, setting the result in redis.
+func (h *handler) fetchMediaID(msg courier.Msg, mimeType, mediaURL string) (string, []*courier.ChannelLog, error) {
+	var logs []*courier.ChannelLog
+
+	// check in cache first
+	rc := h.Backend().RedisPool().Get()
+	defer rc.Close()
+
+	cacheKey := fmt.Sprintf(mediaCacheKeyPattern, msg.Channel().UUID().String())
+	mediaID, err := rcache.Get(rc, cacheKey, mediaURL)
+	if err != nil {
+		return "", logs, errors.Wrapf(err, "error reading media id from redis: %s : %s", cacheKey, mediaURL)
+	} else if mediaID != "" {
+		return mediaID, logs, nil
+	}
+
+	// check in failure cache
+	failKey := fmt.Sprintf("%s-%s", msg.Channel().UUID().String(), mediaURL)
+	found, _ := failedMediaCache.Get(failKey)
+
+	// any non nil value means we cached a failure, don't try again until our cache expires
+	if found != nil {
+		return "", logs, nil
+	}
+
+	// download media
+	req, err := http.NewRequest("GET", mediaURL, nil)
+	if err != nil {
+		return "", logs, errors.Wrapf(err, "error building media request")
+	}
+	rr, err := utils.MakeHTTPRequest(req)
+	log := courier.NewChannelLogFromRR("Fetching media", msg.Channel(), msg.ID(), rr).WithError("error fetching media", err)
+	logs = append(logs, log)
+	if err != nil {
+		failedMediaCache.Set(failKey, true, cache.DefaultExpiration)
+		return "", logs, nil
+	}
+
+	// upload media to WhatsApp
+	baseURL := msg.Channel().StringConfigForKey(courier.ConfigBaseURL, "")
+	url, err := url.Parse(baseURL)
+	if err != nil {
+		return "", logs, errors.Wrapf(err, "invalid base url set for WA channel: %s", baseURL)
+	}
+	dockerMediaURL, _ := url.Parse("/v1/media")
+
+	req, err = http.NewRequest("POST", dockerMediaURL.String(), bytes.NewReader(rr.Body))
+	if err != nil {
+		return "", logs, errors.Wrapf(err, "error building request to media endpoint")
+	}
+	setWhatsAppAuthHeader(&req.Header, msg.Channel())
+	req.Header.Add("Content-Type", http.DetectContentType(rr.Body))
+	rr, err = utils.MakeHTTPRequest(req)
+	log = courier.NewChannelLogFromRR("Uploading media to WhatsApp", msg.Channel(), msg.ID(), rr).WithError("Error uploading media to WhatsApp", err)
+	logs = append(logs, log)
+	if err != nil {
+		failedMediaCache.Set(failKey, true, cache.DefaultExpiration)
+		return "", logs, errors.Wrapf(err, "error uploading media to whatsapp")
+	}
+
+	// take uploaded media id
+	mediaID, err = jsonparser.GetString(rr.Body, "media", "[0]", "id")
+	if err != nil {
+		return "", logs, errors.Wrapf(err, "error reading media id from response")
+	}
+
+	// put in cache
+	err = rcache.Set(rc, cacheKey, mediaURL, mediaID)
+	if err != nil {
+		return "", logs, errors.Wrapf(err, "error setting media id in cache")
+	}
+
+	return mediaID, logs, nil
+}
+
+func sendWhatsAppMsg(msg courier.Msg, sendPath *url.URL, payload interface{}) (string, string, []*courier.ChannelLog, error) {
 	start := time.Now()
 	jsonBody, err := json.Marshal(payload)
 
@@ -604,7 +709,7 @@ func sendWhatsAppMsg(msg courier.Msg, sendPath *url.URL, token string, payload i
 		return "", "", []*courier.ChannelLog{log}, err
 	}
 	req, _ := http.NewRequest(http.MethodPost, sendPath.String(), bytes.NewReader(jsonBody))
-	req.Header = buildWhatsAppRequestHeader(token)
+	req.Header = buildWhatsAppHeaders(msg.Channel())
 	rr, err := utils.MakeHTTPRequest(req)
 	log := courier.NewChannelLogFromRR("Message Sent", msg.Channel(), msg.ID(), rr).WithError("Message Send Error", err)
 	errPayload := &mtErrorPayload{}
@@ -618,7 +723,7 @@ func sendWhatsAppMsg(msg courier.Msg, sendPath *url.URL, token string, payload i
 		}
 		// check contact
 		baseURL := fmt.Sprintf("%s://%s", sendPath.Scheme, sendPath.Host)
-		rrCheck, err := checkWhatsAppContact(baseURL, token, msg.URN())
+		rrCheck, err := checkWhatsAppContact(msg.Channel(), baseURL, msg.URN())
 
 		if rrCheck == nil {
 			elapsed := time.Now().Sub(start)
@@ -673,12 +778,16 @@ func sendWhatsAppMsg(msg courier.Msg, sendPath *url.URL, token string, payload i
 			}
 		}
 		// try send msg again
-		reqRetry, _ := http.NewRequest(http.MethodPost, sendPath.String(), bytes.NewReader(jsonBody))
-		reqRetry.Header = buildWhatsAppRequestHeader(token)
+		reqRetry, err := http.NewRequest(http.MethodPost, sendPath.String(), bytes.NewReader(jsonBody))
+		if err != nil {
+			return "", "", nil, err
+		}
+		reqRetry.Header = buildWhatsAppHeaders(msg.Channel())
 
 		if retryParam != "" {
 			reqRetry.URL.RawQuery = fmt.Sprintf("%s=1", retryParam)
 		}
+
 		rrRetry, err := utils.MakeHTTPRequest(reqRetry)
 		retryLog := courier.NewChannelLogFromRR("Message Sent", msg.Channel(), msg.ID(), rrRetry).WithError("Message Send Error", err)
 
@@ -692,13 +801,23 @@ func sendWhatsAppMsg(msg courier.Msg, sendPath *url.URL, token string, payload i
 	return "", externalID, []*courier.ChannelLog{log}, err
 }
 
-func buildWhatsAppRequestHeader(token string) http.Header {
-	header := http.Header{
-		"Content-Type":  []string{"application/json"},
-		"Accept":        []string{"application/json"},
-		"Authorization": []string{fmt.Sprintf("Bearer %s", token)},
-		"User-Agent":    []string{utils.HTTPUserAgent},
+func setWhatsAppAuthHeader(header *http.Header, channel courier.Channel) {
+	authToken := channel.StringConfigForKey(courier.ConfigAuthToken, "")
+
+	if channel.ChannelType() == channelTypeD3 {
+		header.Set(d3AuthorizationKey, authToken)
+	} else {
+		header.Set("Authorization", fmt.Sprintf("Bearer %s", authToken))
 	}
+}
+
+func buildWhatsAppHeaders(channel courier.Channel) http.Header {
+	header := http.Header{
+		"Content-Type": []string{"application/json"},
+		"Accept":       []string{"application/json"},
+		"User-Agent":   []string{utils.HTTPUserAgent},
+	}
+	setWhatsAppAuthHeader(&header, channel)
 	return header
 }
 
@@ -725,7 +844,7 @@ type mtContactCheckPayload struct {
 	ForceCheck bool     `json:"force_check"`
 }
 
-func checkWhatsAppContact(baseURL string, token string, urn urns.URN) (*utils.RequestResponse, error) {
+func checkWhatsAppContact(channel courier.Channel, baseURL string, urn urns.URN) (*utils.RequestResponse, error) {
 	payload := mtContactCheckPayload{
 		Blocking:   "wait",
 		Contacts:   []string{fmt.Sprintf("+%s", urn.Path())},
@@ -738,7 +857,7 @@ func checkWhatsAppContact(baseURL string, token string, urn urns.URN) (*utils.Re
 	}
 	sendURL := fmt.Sprintf("%s/v1/contacts", baseURL)
 	req, _ := http.NewRequest(http.MethodPost, sendURL, bytes.NewReader(reqBody))
-	req.Header = buildWhatsAppRequestHeader(token)
+	req.Header = buildWhatsAppHeaders(channel)
 	rr, err := utils.MakeHTTPRequest(req)
 
 	if err != nil {
@@ -776,8 +895,12 @@ func (h *handler) getTemplate(msg courier.Msg) (*MsgTemplating, error) {
 	if err != nil {
 		return nil, errors.Wrapf(err, "invalid templating definition")
 	}
+	// check country
+	if templating.Country != "" {
+		templating.Language = fmt.Sprintf("%s_%s", templating.Language, templating.Country)
+	}
 
-	// map our language from iso639-3 to the WA country / iso638-2 pair
+	// map our language from iso639-3_iso3166-2 to the WA country / iso638-2 pair
 	language, found := languageMap[templating.Language]
 	if !found {
 		return nil, fmt.Errorf("unable to find mapping for language: %s", templating.Language)
@@ -797,74 +920,80 @@ type MsgTemplating struct {
 		UUID string `json:"uuid" validate:"required"`
 	} `json:"template" validate:"required,dive"`
 	Language  string   `json:"language" validate:"required"`
+	Country   string   `json:"country"`
 	Variables []string `json:"variables"`
 }
 
-// mapping from iso639-3 to WA language code
+// mapping from iso639-3_iso3166-2 to WA language code
 var languageMap = map[string]string{
-	"afr": "af",    // Afrikaans
-	"sqi": "sq",    // Albanian
-	"ara": "ar",    // Arabic
-	"aze": "az",    // Azerbaijani
-	"ben": "bn",    // Bengali
-	"bul": "bg",    // Bulgarian
-	"cat": "ca",    // Catalan
-	"zho": "zh_CN", // Chinese (CHN)
-	// zh_HK Chinese (HKG) (unsupported, use zh_CN)
-	// zh_TW Chinese (TAI) (unsupported, use zh_CN)
-	"hrv": "hr", //Croatian
-	"ces": "cs", // Czech
-	"dah": "da", // Danish
-	"nld": "nl", // Dutch
-	"eng": "en", // English
-	// en_GB English (UK) (unsupported, use en)
-	// en_US English (US) (unsupported, use en)
-	"est": "et",  // Estonian
-	"fil": "fil", // Filipino
-	"fin": "fi",  // Finnish
-	"fra": "fr",  // French
-	"deu": "de",  // German
-	"ell": "el",  // Greek
-	"gul": "gu",  // Gujarati
-	"enb": "he",  // Hebrew
-	"hin": "hi",  // Hindi
-	"hun": "hu",  // Hungarian
-	"ind": "id",  // Indonesian
-	"gle": "ga",  // Irish
-	"ita": "it",  // Italian
-	"jpn": "ja",  // Japanese
-	"kan": "kn",  // Kannada
-	"kaz": "kk",  // Kazakh
-	"kor": "ko",  // Korean
-	"lao": "lo",  // Lao
-	"jav": "lv",  // Latvian
-	"lit": "lt",  // Lithuanian
-	"mkd": "mk",  // Macedonian
-	"msa": "ms",  // Malay
-	"mar": "mr",  // Marathi
-	"nob": "nb",  // Norwegian
-	"fas": "fa",  // Persian
-	"pol": "pl",  // Polish
-	// "pt_BR" Portuguese (BR)  (unsupported, use pt_PT)
-	"por": "pt_PT", // Portuguese (POR)
-	"pan": "pa",    // Punjabi
-	"ron": "ro",    // Romanian
-	"rus": "ru",    // Russian
-	"srp": "sr",    // Serbian
-	"slk": "sk",    // Slovak
-	"slv": "sl",    // Slovenian
-	"spa": "es",    // Spanish
-	// es_AR Spanish (ARG) (unsupported, use es)
-	// es_ES Spanish (SPA) (unsupported, use es)
-	// es_MX Spanish (MEX) (unsupported, use es)
-	"swa": "sw", // Swahili
-	"swe": "sv", // Swedish
-	"tam": "ta", // Tamil
-	"tel": "te", // Telugu
-	"tha": "th", // Thai
-	"tur": "tr", // Turkish
-	"ukr": "uk", // Ukrainian
-	"urd": "ur", // Urdu
-	"uzb": "uz", // Uzbek
-	"vie": "vi", // Vietnamese
+	"afr":    "af",    // Afrikaans
+	"sqi":    "sq",    // Albanian
+	"ara":    "ar",    // Arabic
+	"aze":    "az",    // Azerbaijani
+	"ben":    "bn",    // Bengali
+	"bul":    "bg",    // Bulgarian
+	"cat":    "ca",    // Catalan
+	"zho":    "zh_CN", // Chinese
+	"zho_CN": "zh_CN", // Chinese (CHN)
+	"zho_HK": "zh_HK", // Chinese (HKG)
+	"zho_TW": "zh_TW", // Chinese (TAI)
+	"hrv":    "hr",    // Croatian
+	"ces":    "cs",    // Czech
+	"dah":    "da",    // Danish
+	"nld":    "nl",    // Dutch
+	"eng":    "en",    // English
+	"eng_GB": "en_GB", // English (UK)
+	"eng_US": "en_US", // English (US)
+	"est":    "et",    // Estonian
+	"fil":    "fil",   // Filipino
+	"fin":    "fi",    // Finnish
+	"fra":    "fr",    // French
+	"deu":    "de",    // German
+	"ell":    "el",    // Greek
+	"gul":    "gu",    // Gujarati
+	"hau":    "ha",    // Hausa
+	"enb":    "he",    // Hebrew
+	"hin":    "hi",    // Hindi
+	"hun":    "hu",    // Hungarian
+	"ind":    "id",    // Indonesian
+	"gle":    "ga",    // Irish
+	"ita":    "it",    // Italian
+	"jpn":    "ja",    // Japanese
+	"kan":    "kn",    // Kannada
+	"kaz":    "kk",    // Kazakh
+	"kor":    "ko",    // Korean
+	"lao":    "lo",    // Lao
+	"lav":    "lv",    // Latvian
+	"lit":    "lt",    // Lithuanian
+	"mal":    "ml",    // Malayalam
+	"mkd":    "mk",    // Macedonian
+	"msa":    "ms",    // Malay
+	"mar":    "mr",    // Marathi
+	"nob":    "nb",    // Norwegian
+	"fas":    "fa",    // Persian
+	"pol":    "pl",    // Polish
+	"por":    "pt_PT", // Portuguese
+	"por_BR": "pt_BR", // Portuguese (BR)
+	"por_PT": "pt_PT", // Portuguese (POR)
+	"pan":    "pa",    // Punjabi
+	"ron":    "ro",    // Romanian
+	"rus":    "ru",    // Russian
+	"srp":    "sr",    // Serbian
+	"slk":    "sk",    // Slovak
+	"slv":    "sl",    // Slovenian
+	"spa":    "es",    // Spanish
+	"spa_AR": "es_AR", // Spanish (ARG)
+	"spa_ES": "es_ES", // Spanish (SPA)
+	"spa_MX": "es_MX", // Spanish (MEX)
+	"swa":    "sw",    // Swahili
+	"swe":    "sv",    // Swedish
+	"tam":    "ta",    // Tamil
+	"tel":    "te",    // Telugu
+	"tha":    "th",    // Thai
+	"tur":    "tr",    // Turkish
+	"ukr":    "uk",    // Ukrainian
+	"urd":    "ur",    // Urdu
+	"uzb":    "uz",    // Uzbek
+	"vie":    "vi",    // Vietnamese
+	"zul":    "zu",    // Zulu
 }